requirements.txt
ansible>=2.7,<3
ansible-lint>=4,<5
docker>=4,<5
flake8>=3,<4
molecule>=3,<4
molecule-docker>=0.2,<1
pytest>=6,<7
pytest-testinfra>=6,<7
